requirement.txt
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
Deprecated==1.2.14
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
pydantic==1.10.12
pydub==0.25.1
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
typing_extensions==4.7.1
urllib3==2.0.4
vonage==3.8.0
vonage-jwt==1.0.0
Werkzeug==2.3.7
wrapt==1.15.0
